Added leaf definitions.
Not sure if this is the right way to go about this.
I'm considering adding some meta characters that specify direction.

Instead of 't' for an UpLeftLeaf, you would write the meta character,
say 'l' for leaf, and then the direction, say 'n' for north.

Considering we will (probably) only ever have Branches and Leaves, our
character set shouldn't ever get large enough for this to be an issue.
Plus it's less cumbersome to parse.